requirements.txt
streamlit
pillow
requests
uuid
spacy
azure.storage.blob
azure.cognitiveservices.vision.computervision
msrest

